Move getLoggedInMentee endpoint to Program Controller
